git-svn-id: svn://svn.code.sf.net/p/d2x-xl/code/trunk@9332 01821946-851b-0410-b7f4-f1477c70e57a

+++ gameio/mission.cpp
@@ -293,7 +293,7 @@ char* Trim (char* s)
 if (s) {
 	while (::isspace (*s))
 		s++;
-	for (int i = strlen (s); --i; ) {
+	for (int i = (int) strlen (s); --i; ) {
 		if (::isspace (s [i]))
 			s [i] = '\0';
 		else
